requirements.txt
numpy
cvxpy
pytorch
